[charts-premium] Share helper function
